Move utils containers to components
