+++ client/src/views/sources/index.tsx
@@ -5,7 +5,7 @@ import { Mutation, Query, Subscription } from "react-apollo";
 import { connect } from "react-redux";
 import { withRouter } from "react-router-dom";
 
-import Navbar from "../../utils/navbar";
+import Navbar from "../../components/navbar";
 
 import { changeSelectedSource } from "../../actions/selectedSource";
